+++ src/components/dialog/ha-paper-dialog.ts
@@ -1,6 +1,8 @@
 import "@polymer/paper-dialog/paper-dialog";
 import { mixinBehaviors } from "@polymer/polymer/lib/legacy/class";
 import { HaIronFocusablesHelper } from "./ha-iron-focusables-helper.js";
+// tslint:disable-next-line
+import { PaperDialogElement } from "@polymer/paper-dialog/paper-dialog";
 
 const paperDialogClass = customElements.get("paper-dialog");
 
@@ -13,10 +15,10 @@ const haTabFixBehaviorImpl = {
 
 // paper-dialog that uses the haTabFixBehaviorImpl behvaior
 // export class HaPaperDialog extends paperDialogClass {}
-export class HaPaperDialog extends mixinBehaviors(
-  [haTabFixBehaviorImpl],
-  paperDialogClass
-) {}
+// @ts-ignore
+export class HaPaperDialog
+  extends mixinBehaviors([haTabFixBehaviorImpl], paperDialogClass)
+  implements PaperDialogElement {}
 
 declare global {
   interface HTMLElementTagNameMap {
