+++ src/dialogs/config-flow/dialog-config-flow.ts
@@ -16,10 +16,10 @@ import "@polymer/paper-spinner/paper-spinner";
 import "../../components/ha-form";
 import "../../components/ha-markdown";
 import "../../resources/ha-style";
-import "../../panels/lovelace/components/dialog/ha-dialog";
+import "../../components/dialog/ha-paper-dialog";
 // Not duplicate, is for typing
 // tslint:disable-next-line
-import { HaPaperDialog } from "../../panels/lovelace/components/dialog/ha-dialog";
+import { HaPaperDialog } from "../../components/dialog/ha-paper-dialog";
 import { haStyleDialog } from "../../resources/styles";
 import {
   fetchConfigFlow,
