+++ src/panels/config/users/ha-dialog-add-user.js
@@ -3,7 +3,7 @@ import "@polymer/paper-spinner/paper-spinner";
 import { html } from "@polymer/polymer/lib/utils/html-tag";
 import { PolymerElement } from "@polymer/polymer/polymer-element";
 
-import "../../lovelace/components/dialog/ha-dialog";
+import "../../../components/dialog/ha-paper-dialog";
 import "../../../resources/ha-style";
 
 import LocalizeMixin from "../../../mixins/localize-mixin";
