+++ src/panels/lovelace/editor/hui-dialog-save-config.ts
@@ -8,9 +8,9 @@ import {
   property,
 } from "lit-element";
 import "@polymer/paper-spinner/paper-spinner";
-import "../components/dialog/ha-dialog";
+import "../../../components/dialog/ha-paper-dialog";
 // tslint:disable-next-line:no-duplicate-imports
-import { HaPaperDialog } from "../components/dialog/ha-dialog";
+import { HaPaperDialog } from "../../../components/dialog/ha-paper-dialog";
 import "@material/mwc-button";
 
 import { haStyleDialog } from "../../../resources/styles";
